+++ inc/types/growing_point.hpp
@@ -8,7 +8,7 @@
  * @brief Coordinates representing the corner of the block waiting to be encoded/decoded.
  */
 typedef struct Growing_point {
-  size_t x; ///< x coordinate
-  size_t y; ///< y coordinate
-  Growing_point(size_t _x, size_t _y): x(_x), y(_y){};
+  uint16_t x; ///< x coordinate
+  uint16_t y; ///< y coordinate
+  Growing_point(uint16_t _x, uint16_t _y): x(_x), y(_y){};
 } Growing_point;
